+++ src-tauri/src/main.rs
@@ -2,6 +2,6 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
-    dsh_launcher_lib::run();
+    dsh_desktop_lib::run();
     std::process::exit(0);
 }